[백엔드 :: HTTP 통신 합의된 내용으로 동기화]

+++ src/main/java/com/example/oose_23_3_back/historycontrol/entity/PaymentDetail.java
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 public class PaymentDetail {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer paymentDetailNumber;
+    private Long paymentDetailNumber;
     @ManyToOne
     private Member member;
     @ManyToOne
