+++ src/main/java/com/example/oose_23_3_back/managementcontrol/repository/RentalOfficeRepository.java
@@ -3,5 +3,5 @@ package com.example.oose_23_3_back.managementcontrol.repository;
 import com.example.oose_23_3_back.managementcontrol.entity.RentalOffice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalOfficeRepository extends JpaRepository<RentalOffice, Integer> {
+public interface RentalOfficeRepository extends JpaRepository<RentalOffice, Long> {
 }
